requirements.txt
pypdf>=5.0.0
paramiko>=3.0.0
psycopg2-binary>=2.9.0